fix: store context id as int64_t
Ensuring that it doesn't wrap easily with a large number
of context creation on some malformed web pages.

+++ atom/renderer/renderer_client_base.h
@@ -58,7 +58,7 @@ class RendererClientBase : public content::ContentRendererClient {
   bool isolated_world_;
   std::string renderer_client_id_;
   // An increasing ID used for indentifying an V8 context in this process.
-  int next_context_id_ = 0;
+  int64_t next_context_id_ = 0;
 };
 
 }  // namespace atom
